Clear Works
There is an issue with how I am checking the focus.  Works fine for ContextMenu, but if you select, hover over a different tasview, then use the Menu at the top of the form then it can select the wrong tasview to do an edit. Try MouseUp or a SelectedItem event to set a variable for focused tasview.