Stop exposing PicoDrive as a selectable option for Genesis
The core as it is is a very poor choice for Genesis. Much worse compatibility, very noticeably worse audio, and it's not like GPGX is hurting badly in the performance department. It's also very outdated compared to upstream (making these problems much worse than if it was updated). It was only in BizHawk to handle 32X games to begin with (as there aren't many options for 32X to begin with), with Genesis fallback support (without user selectability!) just tacked on "because we can."

Perhaps with a core update it could be more reasonable to allow users to select this, but for now, nope.

Closes #4251, #4250, #4235