fix tests issue
It is the memory callback that is able to return null or an uint. Before this commit, the code relied on Callback to return either null or an uint, and then MemoryCallback to return the final uint, the value sent by the core or the override value, but in the end an uint always reached Call. The next commit will be a refactor on NamedLuaFunction to fully acommodate to tests (this commit fixes the tests errors, Bizhawk does build, but the actual desired functionality won't work)