dsda: allow full byte range in BOTH directions for frac turn
since they're split and we can't set hi byte to -0, there was no way to turn below -1. [-256, 256] is not an accurate range per se, but due to how things are added on the managed side before sending inputs to the core, it feels like the most useful one. for shorttics one would only have access to hi byte, and with longtics one can still simulate that behavior - by only controlling said hi byte! and only introduce "sub-angle" when it's actually needed for something specific