dsda: multi-wad hashes
just like mame we're hashing the wad files, concating them, hashing that, and the result goes to movie header. for singular IWADs loaded on their own just the actual hash is used instead, to make it clear to users.

printing hashes of all files to header might be cool, but I'm not sure it's warranted - usually there's only 1 PWAD