refactor ProcessInput to simplify logic and remove the distinction "internal hotkeys"; all hotkeys go to the hotkey controller now even if we don't expect to read that hotkey from the controller.
Fix: Hotkey overrides input would cause an input to continue after a button was released if the input is bound to a combination such as Shift+A and an "internal hotkey" is bound to a single button (such as A). Also input bound to Alt would not have the alt ignored by menus.
Passes tests HotkeyOverrideDoesNotEatReleaseEvents and InputIsNotSeenAsUnbound