Add Analyzer to ban `override bool Equals(object?)` on `ref struct`s
It's pointless to call this method, and when you omit it the compiler
tries to emit a call to the base method which is a simple build error.
Also ban overriding `GetHashCode`, because I don't think that's useful
outside of hash tables (which you obvously can't use ref structs in),
and implementing one without the other is a bad practice anyway and will
be picked up by another Analyzer.