Store the absolute timestamp separately in RawTOCEntry
This is more a "quick fix" for DVD support if anything, and more fixes will be needed for various parsers to handle DVDs.

DVDs aren't like CDs. They can be seen as being like a CD without subcode (and thus single tracked, excluding leadin and leadout) and only 2048 bytes per sector, so similar to a simple single tracked Mode 1 CD, and thus they more or less get parsed as such for various formats.

However, DVDs can be rather large. They can be large enough, in fact, the absolute timestamp in subcode (which is used for raw TOC entries and eventually the parsed track LBAs) cannot hold the timestamp, due to the constraints of BCD (or rather subcode BCD, since it's only a byte large). In practice, this means if a DVD is over 719999 sectors (i.e. ~1406MiBs), it will end up reporting a leadout timestamp of >=160:00:00. 160 minutes cannot be held in BCD (that's 0x100, larger than a byte), so it ends up "wrapping around" to 0 (and so on with later values). This is bad, as the upmost synth provider relies on the leadout LBA (and thus leadout timestamp in raw TOC entries) to be correct in order to determine if a leadout sector should be synthed. Cores also generally will use the leadout LBA to know the length of the disc. So, if the leadout LBA is wrong, the entire system will break down. Storing the absolute timestamp separately ensures this won't happen, so a core wanting a DVD should mostly function fine, even for large DVDs.

Note, of course, various formats, such as (in particular) CHD, can't be parsed correctly due to such formats explicitly defining DVDs in them. This will be fixed later.