Revert "Remove default bindings for keyboards (resolves #3257)"
This reverts commit ec3a23259cf0f9528aef661b9df51846997485f4.

We're adding more keyboard based systems now (most notably DOSBox-X core) so it'd be a pain to keep forcing users to bind the entire keyboard for every release. Cores must be usable out of the box, first and foremost for casual play. Sure we're a TASing emulator, but we're also a casual gaming emulator. A rather popular at that, and we want the experience to be much nicer than other multisystem emulators that aim at casual gamers. Not requiring any obscure configuration just to be able to run a game is an important part of this.