Partially revert 9df29c036eea4ecca9f9ee632a65bf7b09549832
`ArraySegment` is just like a Span, in that it just holds a reference to an array with an offset and count. So this code just ended up making a list of ArraySegments, nearly all of which ended up referencing the same array which is constantly mutated here, which gave garbage results when hashing ultimately occurred.