dsda: support using doom2 as firmware
fixes #4234

overwhelming majority of PWADs rely on doom2, so in most cases nothing else will be needed. pointing to it once and then using it automatically feels like an improvement even over upstream.

if IWAD is loaded normally, we don't try to look for the firmware one. so loading just an IWAD or an IWAD+PWAD within xml, will not trigger the failure/clash. loading the wrong PWAD with firmware doom2 will result in whatever is already happening if you do that with xml.

ability to load other IWADs like this is low priority due to rarity.