workaround for cores showing internal load errors with xml loader
if we pass the core's own error message over to the user without xml, the dialog they see is the regular "core accepted the rom + error message" one. but if the same happens for something loaded through xml, the dialog is instead "no core could load this + 2 copies of the same error/stack".

turns out there's a special case when MissingFirmwareException is thrown to avoid that second dialog. but internal errors are not MissingFirmwareException, so I added my own and a check for it.

unfortunately when running the debug build of hawk PLUS release build of dsda and causing such an internal error, there's still stack overflow after the throw...